:bug: HF_0000034: [156] CDIS to Catapult Data for KDS

+++ database/migrations/2026_04_10_144923_add_prep_time_column_in_product_uom_and_kitchen_item_setup_table.php
@@ -43,7 +43,7 @@ class AddPrepTimeColumnInProductUomAndKitchenItemSetupTable extends Migration
                     }
 
                     if (! Schema::hasColumn($alterTable[0], 'max_prep_time')) {
-                        $table->integer('max_prep_time')->default(0)->after($alterTable[1]);
+                        $table->decimal('max_prep_time', 23, 6)->default(0.000000)->after($alterTable[1]);
                     }
                 });
             }
